feat(hotkeys): introduce EditModeProvider and refactor shortcut handling
- Added `EditModeProvider` to manage edit mode state, enhancing shortcut behavior across the application.
- Updated components and hooks to utilize the new provider, ensuring consistent shortcut functionality in edit mode.
- Refactored shortcut imports and definitions to streamline the handling of hotkeys, improving code organization.
- Introduced new hooks for managing app hotkeys and sequences, enhancing keyboard navigation capabilities.
- Added tests for the new hotkey functionality and EditModeProvider to ensure reliability and correctness.
