+++ packages/web/src/hotkeys/constants/hotkey.constants.ts
@@ -1,0 +1,1 @@
+export const EDIT_MODE_TIMEOUT_MS = 1000;
