+++ packages/web/src/hotkeys/registry/shortcut.registry.ts
