+++ packages/web/src/views/Calendar/components/Header/Reminder/Reminder.tsx
@@ -10,11 +10,11 @@ import {
 } from "react";
 import { STORAGE_KEYS } from "@web/common/constants/storage.constants";
 import { ID_REMINDER_INPUT } from "@web/common/constants/web.constants";
-import { useAppHotkey } from "@web/common/hooks/useAppHotkey";
 import { theme } from "@web/common/styles/theme";
 import { TooltipWrapper } from "@web/components/Tooltip/TooltipWrapper";
 import { selectReminder } from "@web/ducks/events/selectors/view.selectors";
 import { viewSlice } from "@web/ducks/events/slices/view.slice";
+import { useAppHotkey } from "@web/hotkeys/hooks/useAppHotkey";
 import { useAppDispatch, useAppSelector } from "@web/store/store.hooks";
 import { generateHandDrawnUnderline } from "@web/views/Calendar/components/Header/Reminder/reminder-util";
 import {
