+++ packages/web/src/views/Day/components/DayCmdPalette.tsx
@@ -5,7 +5,6 @@ import dayjs from "@core/util/date/dayjs";
 import { useConnectGoogle } from "@web/auth/google/hooks/useConnectGoogle/useConnectGoogle";
 import { moreCommandPaletteItems } from "@web/common/constants/more.cmd.constants";
 import { useAuthCmdItems } from "@web/common/hooks/useAuthCmdItems";
-import { SHORTCUTS } from "@web/common/shortcuts/shortcut.registry";
 import { pressKey } from "@web/common/utils/dom/event-emitter.util";
 import {
   openEventFormCreateEvent,
@@ -13,6 +12,7 @@ import {
 } from "@web/common/utils/event/event.util";
 import { selectIsCmdPaletteOpen } from "@web/ducks/settings/selectors/settings.selectors";
 import { settingsSlice } from "@web/ducks/settings/slices/settings.slice";
+import { SHORTCUTS } from "@web/hotkeys/registry/shortcut.registry";
 import { useAppDispatch, useAppSelector } from "@web/store/store.hooks";
 
 interface DayCmdPaletteProps {
