+++ packages/web/src/views/Day/components/Toasts/UndoToast/UndoDeleteToast.test.tsx
@@ -1,7 +1,7 @@
 import { toast } from "react-toastify";
 import "@testing-library/jest-dom";
 import { fireEvent, render, screen } from "@testing-library/react";
-import { getModifierKeyTestId } from "@web/common/utils/shortcut/shortcut.util";
+import { getModifierKeyTestId } from "@web/hotkeys/util/shortcut.util";
 import {
   UndoDeleteToast,
   showUndoDeleteToast,
