+++ packages/web/src/views/Forms/EventForm/DuplicateMenuButton.tsx
@@ -1,7 +1,7 @@
 import type React from "react";
 import { Copy } from "@phosphor-icons/react";
-import { getModifierKeyIcon } from "@web/common/utils/shortcut/shortcut.util";
 import { Text } from "@web/components/Text";
+import { getModifierKeyIcon } from "@web/hotkeys/util/shortcut.util";
 import MenuItem from "@web/views/Forms/ActionsMenu/MenuItem";
 
 interface Props {
