+++ packages/web/src/views/Forms/EventForm/SaveSection/SaveSection.tsx
@@ -1,10 +1,10 @@
 import type React from "react";
 import { useCallback } from "react";
 import { type Priority } from "@core/constants/core.constants";
-import { getModifierKeyIcon } from "@web/common/utils/shortcut/shortcut.util";
 import { Btn, StyledSaveBtn } from "@web/components/Button/styled";
 import { Text } from "@web/components/Text";
 import { TooltipWrapper } from "@web/components/Tooltip/TooltipWrapper";
+import { getModifierKeyIcon } from "@web/hotkeys/util/shortcut.util";
 import { StyledSubmitRow } from "@web/views/Forms/EventForm/styled";
 
 interface Props {
