+++ packages/web/src/views/Now/view/NowView.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { usePointerPosition } from "@web/common/hooks/usePointerPosition";
 import { useSidebarState } from "@web/common/hooks/useSidebarState";
-import { getShortcuts } from "@web/common/utils/shortcut/data/shortcuts.data";
+import { getShortcuts } from "@web/hotkeys/registry/shortcuts.data";
 import { StyledCalendar } from "@web/views/Calendar/styled";
 import { Header } from "@web/views/Day/components/Header/Header";
 import { ShortcutsSidebar } from "@web/views/Day/components/ShortcutsSidebar/ShortcutsSidebar";
